FIX: Hide TOC header when no TOC
The LoveIt theme we are modifying would include the auto TOC even if
there were no headings from which to create a TOC. This would leave an
awkward title with nothing under it.  Here we remove this title, but
retain the parent dom node to maintain consistent spacing with other
pages.